requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.3.2
faiss-cpu==1.8.0.post1
pymupdf==1.24.10
arxiv==2.1.3
structlog==24.4.0
sentence-transformers==3.2.1
pytest==8.3.3
